Mere af project lavet.

+++ EksamensprojektVinter2021/src/main/java/com/example/eksamensprojektvinter2021/Models/Project.java
@@ -83,7 +83,7 @@ public class Project {
         this.totalTime = totalTime;
     }
 
-    public double getTotalTime() {
+    public int getTotalTime() {
         return totalTime;
     }
